fix: adjust the code to separate RedisCache
