+++ src/main/java/com/achobeta/util/RedisBloomFilter.java
@@ -1,4 +1,4 @@
-package com.achobeta.domain.shortlink.component;
+package com.achobeta.util;
 
 import lombok.RequiredArgsConstructor;
 import org.assertj.core.util.Preconditions;
@@ -9,9 +9,6 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class RedisBloomFilter {
 
-    private static final String BLOOM_FILTER_NAME = "LINK-CODE-LIST";
-
-
     private final RedisTemplate redisTemplate;
 
     private BloomFilterHelper bloomFilterHelper = new BloomFilterHelper();
@@ -19,23 +16,23 @@ public class RedisBloomFilter {
     /**
      * 根据给定的布隆过滤器添加值
      */
-    public <T> void add(T value) {
+    public <T> void add(String bloomFilterName, T value) {
         Preconditions.checkArgument(bloomFilterHelper != null, "bloomFilterHelper不能为空");
         int[] offset = bloomFilterHelper.murmurHashOffset(value);
         for (int i : offset) {
             // todo: 设置超时时间
-            redisTemplate.opsForValue().setBit(BLOOM_FILTER_NAME, i, true);
+            redisTemplate.opsForValue().setBit(bloomFilterName, i, true);
         }
     }
 
     /**
      * 根据给定的布隆过滤器判断值是否存在
      */
-    public <T> boolean contains(T value) {
+    public <T> boolean contains(String bloomFilterName, T value) {
         Preconditions.checkArgument(bloomFilterHelper != null, "bloomFilterHelper不能为空");
         int[] offset = bloomFilterHelper.murmurHashOffset(value);
         for (int i : offset) {
-            if (!redisTemplate.opsForValue().getBit(BLOOM_FILTER_NAME, i)) {
+            if (!redisTemplate.opsForValue().getBit(bloomFilterName, i)) {
                 return false;
             }
         }
