+++ src/main/java/com/achobeta/util/RedisCache.java
@@ -1,8 +1,7 @@
-package com.achobeta.domain.shortlink.component;
+package com.achobeta.util;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.sf.jsqlparser.statement.select.KSQLWindow;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
 
@@ -29,6 +28,12 @@ public class RedisCache {
         redisTemplate.opsForValue().set(key, value);
     }
 
+    /**
+     * 获取键值
+     * @param key 键
+     * @return 键对应的值，并封装成 Optional 对象
+     * @param <T>
+     */
     public <T> Optional<T> getCacheObject(final String key) {
         T value = (T) redisTemplate.opsForValue().get(key);
         log.info("查询Redis\t[{}]-[{}]", key, value);
@@ -37,7 +42,6 @@ public class RedisCache {
  
     /**
      * 删除单个对象
-     *
      * @param key
      */
     public boolean deleteObject(final String key) {
@@ -47,22 +51,23 @@ public class RedisCache {
 
     /**
      * 加入布隆过滤器
-     * @param key 键值
-     * @return
+     * @param bloomFilterName 隆过滤器的名字
+     * @param key key 键
      */
-    public void addToBloomFilter(final String key) {
-        log.info("加入布隆过滤器\tkey[{}]", key);
-        redisBloomFilter.add(key);
+    public void addToBloomFilter(final String bloomFilterName, final String key) {
+        log.info("加入布隆过滤器[{}]\tkey[{}]", bloomFilterName, key);
+        redisBloomFilter.add(bloomFilterName, key);
     }
 
     /**
      * 布隆过滤器是否存在该键值
-     * @param key 键值
-     * @return
+     * @param bloomFilterName 布隆过滤器的名字
+     * @param key 键
+     * @return 键是否存在
      */
-    public boolean containsInBloomFilter(final String key) {
-        boolean flag = redisBloomFilter.contains(key);
-        log.info("key[{}]\t是否存在于布隆过滤器:\t{}", key, flag);
+    public boolean containsInBloomFilter(final String bloomFilterName, final String key) {
+        boolean flag = redisBloomFilter.contains(bloomFilterName, key);
+        log.info("key[{}]\t是否存在于布隆过滤器[{}]:\t{}", key, bloomFilterName, flag);
         return flag;
     }
